+++ packages/migrations/src/ShopSys/MigrationBundle/Command/MigrateCommand.php
@@ -2,11 +2,16 @@
 
 namespace ShopSys\MigrationBundle\Command;
 
+use Doctrine\DBAL\Migrations\Configuration\Configuration;
+use Doctrine\DBAL\Migrations\OutputWriter;
+use Doctrine\DBAL\Migrations\Tools\Console\Helper\ConfigurationHelper;
 use Doctrine\ORM\EntityManagerInterface;
+use ShopSys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class MigrateCommand extends Command
 {
@@ -24,11 +29,28 @@ class MigrateCommand extends Command
     private $em;
 
     /**
-     * @param \Doctrine\ORM\EntityManagerInterface $em
+     * @var \Symfony\Component\DependencyInjection\ContainerInterface
      */
-    public function __construct(EntityManagerInterface $em)
-    {
+    private $container;
+
+    /**
+     * @var \ShopSys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator
+     */
+    private $migrationsLocator;
+
+    /**
+     * @param \Doctrine\ORM\EntityManagerInterface $em
+     * @param \Symfony\Component\DependencyInjection\ContainerInterface $container
+     * @param \ShopSys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator $migrationsLocator
+     */
+    public function __construct(
+        EntityManagerInterface $em,
+        ContainerInterface $container,
+        MigrationsLocator $migrationsLocator
+    ) {
         $this->em = $em;
+        $this->container = $container;
+        $this->migrationsLocator = $migrationsLocator;
 
         parent::__construct();
     }
@@ -48,18 +70,48 @@ class MigrateCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         try {
-            $this->em->transactional(function () use ($output) {
-                $this->executeDoctrineMigrateCommand($output);
+            $migrationsConfiguration = $this->createAndRegisterMigrationsConfiguration($output);
 
-                $output->writeln('');
+            foreach ($this->migrationsLocator->getMigrationsLocations() as $migrationsLocation) {
+                $migrationsConfiguration->setMigrationsDirectory($migrationsLocation->getDirectory());
+                $migrationsConfiguration->setMigrationsNamespace($migrationsLocation->getNamespace());
 
-                $this->executeCheckSchemaCommand($output);
-            });
+                $output->writeln('Installing migrations from ' . $migrationsLocation->getDirectory() . ' in namespace ' . $migrationsLocation->getNamespace());
+
+                $this->em->transactional(function () use ($output) {
+                    $this->executeDoctrineMigrateCommand($output);
+
+                    $output->writeln('');
+                });
+            }
+
+            $output->writeln('Migrations from all sources has been installed.');
+
+            $this->executeCheckSchemaCommand($output);
         } catch (\Exception $ex) {
-            $message = "Database migration process did not run properly. Transaction was reverted.\n"
+            $message = "Database migration process did not run properly. Last transaction was reverted.\n"
                 . 'For more informations see the previous exception.';
             throw new \ShopSys\MigrationBundle\Command\Exception\MigrateCommandException($message, $ex);
         }
+    }
+
+    /**
+     * @param \Symfony\Component\Console\Output\OutputInterface $output
+     * @return \Doctrine\DBAL\Migrations\Configuration\Configuration
+     */
+    private function createAndRegisterMigrationsConfiguration(OutputInterface $output)
+    {
+        $outputWriter = new OutputWriter(
+            function ($message) use ($output) {
+                return $output->writeln($message);
+            }
+        );
+
+        $migrationsConfiguration = new Configuration($this->em->getConnection(), $outputWriter);
+        $configurationHelper = new ConfigurationHelper($this->em->getConnection(), $migrationsConfiguration);
+        $this->getApplication()->getHelperSet()->set($configurationHelper, 'configuration');
+
+        return $migrationsConfiguration;
     }
 
     /**
@@ -72,6 +124,7 @@ class MigrateCommand extends Command
             'command' => 'doctrine:migrations:migrate',
             '--allow-no-migration' => true,
         ];
+
         $input = new ArrayInput($arguments);
         $input->setInteractive(false);